+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [[Unreleased]](https://github.com/kovihome/ReqSmith)
 ### Added
+- New ReqM element definition: style
+- StdRepo: Style feature
 
 ### Modified